+++ apps/frontend/src/app/pages/assignment-problem/assignment-problem.module.ts
@@ -4,8 +4,7 @@ import {MatButtonModule} from '@angular/material/button';
 import {MatSelectModule} from '@angular/material/select';
 
 import {InfoCardModule} from '@components/info-card/info-card.module';
-import {InputTableModule} from '@components/input-table/input-table.module';
-import {SimpleTableModule} from '@components/simple-table/simple-table.module';
+import {InputTableModule, SimpleTableModule} from '@opres/generatable-tables';
 import {TranslateModule} from '@ngx-translate/core';
 
 import {AssignmentProblemPageComponent} from './assignment-problem.page';