+++ apps/frontend/src/app/pages/transport-problem/+table/transport-table.component.spec.ts
@@ -4,11 +4,10 @@ import {ComponentFixture, TestBed} from '@angular/core/testing';
 import {MatButtonHarness} from '@angular/material/button/testing';
 import {NoopAnimationsModule} from '@angular/platform-browser/animations';
 
+import {InputTableService} from '@opres/generatable-tables';
 import {TranslateModule} from '@ngx-translate/core';
 
-import {InputTableService} from '../../../components/input-table/input-table.service';
 import {TransportProblemModule} from '../transport-problem.module';
-import {TransportProblemService} from '../transport-problem.service';
 
 import {TransportTableComponent} from './transport-table.component';
 
@@ -16,7 +15,6 @@ describe('TransportTableComponent', () => {
   let fixture: ComponentFixture<TransportTableComponent>;
   let component: TransportTableComponent;
   let loader: HarnessLoader;
-  let transportProblemService: TransportProblemService;
   let inputTableService: InputTableService;
 
   beforeEach(async () => {
@@ -31,7 +29,6 @@ describe('TransportTableComponent', () => {
     fixture = TestBed.createComponent(TransportTableComponent);
     component = fixture.componentInstance;
     loader = TestbedHarnessEnvironment.loader(fixture);
-    transportProblemService = TestBed.inject(TransportProblemService);
     inputTableService = TestBed.inject(InputTableService);
     fixture.detectChanges();
   });
@@ -40,16 +37,12 @@ describe('TransportTableComponent', () => {
     expect(component).toBeInstanceOf(TransportTableComponent));
 
   it('should check default values', () => {
-    expect(component.storages$.getValue()).toEqual(4);
-    expect(component.shops$.getValue()).toEqual(4);
+    expect(component.storagesCount).toEqual(4);
+    expect(component.shopsCount).toEqual(4);
   });
 
   it('should clear the tables with the button', async () => {
     const inputTableServiceClearSpy = jest.spyOn(inputTableService, 'clear');
-    const transportProblemServiceClearSpy = jest.spyOn(
-      transportProblemService,
-      'clear',
-    );
     const tableClearEmitSpy = jest.spyOn(component.tableClear, 'emit');
     const clearButton = await loader.getHarness(
       MatButtonHarness.with({selector: '[data-test-id="table-clear-button"]'}),
@@ -61,7 +54,6 @@ describe('TransportTableComponent', () => {
       'storageStocks',
       'shopDemands',
     ]);
-    expect(transportProblemServiceClearSpy).toHaveBeenCalled();
     expect(tableClearEmitSpy).toHaveBeenCalled();
   });
 });