+++ apps/frontend/src/app/pages/transport-problem/+table/transport-table.component.ts
@@ -2,14 +2,12 @@ import {
   ChangeDetectionStrategy,
   Component,
   EventEmitter,
+  Input,
   Output,
 } from '@angular/core';
 
-import {InputTableService} from '@components/input-table/input-table.service';
-import {Table} from '@shared/types/table.types';
-
-import {TransportProblemService} from '../transport-problem.service';
-import {Demands, Stocks} from '../transport-problem.types';
+import {InputTableService, Table} from '@opres/generatable-tables';
+import {Demands, Stocks} from '@opres/shared-interfaces';
 
 @Component({
   selector: 'transport-table',
@@ -30,9 +28,13 @@ import {Demands, Stocks} from '../transport-problem.types';
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class TransportTableComponent {
-  @Output() tableClear = new EventEmitter<void>();
-  public readonly shops$ = this.transportProblemService.shops$;
-  public readonly storages$ = this.transportProblemService.storages$;
+  @Input() public shopsCount = 4;
+  @Input() public storagesCount = 4;
+
+  @Output() public tableClear = new EventEmitter<void>();
+  @Output() public costChange = new EventEmitter<Table>();
+  @Output() public demandChange = new EventEmitter<Demands>();
+  @Output() public stockChange = new EventEmitter<Stocks>();
 
   public mockCosts: Table = [
     {'0': 8, '1': 7, '2': 3, '3': 2},
@@ -43,28 +45,26 @@ export class TransportTableComponent {
   public mockDemands: Table = [{'0': 18, '1': 32, '2': 35, '3': 20}];
   public mockStocks: Table = [{'0': 15}, {'0': 43}, {'0': 28}, {'0': 19}];
 
-  constructor(
-    private transportProblemService: TransportProblemService,
-    private inputTableService: InputTableService,
-  ) {}
-
-  public clearTables(): void {
-    this.inputTableService.clear(['costs', 'storageStocks', 'shopDemands']);
-    this.transportProblemService.clear();
-    this.tableClear.emit();
-  }
+  constructor(private inputTableService: InputTableService) {}
 
   public onCostChange(values: Table): void {
-    this.transportProblemService.setCosts(values);
+    this.costChange.emit(values);
   }
 
-  public onShopChange(values: Table): void {
-    const arrayOfValues = Object.values(values[0]) as Demands;
-    this.transportProblemService.setShopDemands(arrayOfValues);
+  public onDemandChange(values: Table): void {
+    // convert Table to Demands
+    const arrayOfDemandValues = Object.values(values[0]) as Demands;
+    this.demandChange.emit(arrayOfDemandValues);
   }
 
-  public onStorageChange(values: Table): void {
-    const arrayOfValues = values.map((value) => value['0']) as Stocks;
-    this.transportProblemService.setStorageStocks(arrayOfValues);
+  public onStockChange(values: Table): void {
+    // convert Table to Stocks
+    const arrayOfStockValues = values.map((value) => value['0']) as Stocks;
+    this.stockChange.emit(arrayOfStockValues);
+  }
+
+  public onTablesClear(): void {
+    this.inputTableService.clear(['costs', 'storageStocks', 'shopDemands']);
+    this.tableClear.emit();
   }
 }