+++ apps/frontend/src/app/pages/transport-problem/services/first-phase/vogel-korda.method.service.ts
@@ -1,0 +1,11 @@
+import {Injectable} from '@angular/core';
+
+import {CalculationProcess, TPData} from '@opres/shared-interfaces';
+
+@Injectable()
+export class VogelKordaMethodService {
+  public calculate(transportProblemData: TPData): Array<CalculationProcess> {
+    // todo: wip
+    return [];
+  }
+}