+++ apps/frontend/src/app/pages/transport-problem/transport-problem.constant.ts
@@ -1,9 +1,3 @@
-export const EMPTY_CALCULATION_PROCESS = {
-  transportation: [],
-  demands: [],
-  stocks: [],
-};
-
 export const EMPTY_TP_DATA = {
   costs: [],
   shopDemands: [],