+++ apps/frontend/src/app/pages/transport-problem/transport-problem.service.ts
@@ -1,111 +1,36 @@
 import {Injectable} from '@angular/core';
 
-import {sum} from '@shared/helpers/array.helper';
-import {Row, Table} from '@shared/types/table.types';
-import {BehaviorSubject} from 'rxjs';
-
-import {
-  EMPTY_CALCULATION_PROCESS,
-  EMPTY_TP_DATA,
-} from './transport-problem.constant';
 import {
   CalculationProcess,
-  Demands,
-  Result,
-  Stocks,
   TPData,
-  TransportRow,
+  TPMethods,
   TransportTable,
-} from './transport-problem.types';
+} from '@opres/shared-interfaces';
+import {Observable, of} from 'rxjs';
+
+import {NorthWestMethodService} from './services/first-phase/north-west.method.service';
+import {TableMinimumMethodService} from './services/first-phase/table-minimum.method.service';
+import {VogelKordaMethodService} from './services/first-phase/vogel-korda.method.service';
 
 @Injectable()
 export class TransportProblemService {
-  /** A shop is the equivalent of a column. */
-  public shops$ = new BehaviorSubject<number>(4);
-  /** A storage is the equivalent of a row. */
-  public storages$ = new BehaviorSubject<number>(4);
-  private calculationProcess$ = new BehaviorSubject<CalculationProcess>(
-    EMPTY_CALCULATION_PROCESS,
-  );
-  /** It contains all table data what are necessary for calculations (costs, demands, stocks). */
-  private tpData$ = new BehaviorSubject<TPData>(EMPTY_TP_DATA);
+  private firstPhase = {
+    'north-west': this.northWestMethod,
+    'table-min': this.tableMinimumMethod,
+    'vogel-korda': this.vogelKordaMethod,
+  };
 
-  public get calculationProcess() {
-    return this.calculationProcess$.asObservable();
-  }
+  constructor(
+    private northWestMethod: NorthWestMethodService,
+    private tableMinimumMethod: TableMinimumMethodService,
+    private vogelKordaMethod: VogelKordaMethodService,
+  ) {}
 
-  public setCosts(costs: Table): void {
-    this.tpData$.next({...this.tpData$.getValue(), costs});
-  }
-
-  public setShopDemands(shopDemands: Demands): void {
-    this.tpData$.next({...this.tpData$.getValue(), shopDemands});
-  }
-
-  public setStorageStocks(storageStocks: Stocks): void {
-    this.tpData$.next({...this.tpData$.getValue(), storageStocks});
-  }
-
-  public clear(): void {
-    this.tpData$.next(EMPTY_TP_DATA);
-  }
-
-  public reset(): void {
-    this.calculationProcess$.next(EMPTY_CALCULATION_PROCESS);
-  }
-
-  public northWest(transportProblemData?: TPData): Result {
-    const tpData = transportProblemData || this.tpData$.getValue();
-    if (!this.checkSolvability(tpData))
-      throw new Error('The given problem is not solvable! Try another one.');
-    const {costs} = tpData;
-    const stocks = [...tpData.storageStocks];
-    const demands = [...tpData.shopDemands];
-    const resultTable: TransportTable = this.createResultTableFrom(costs);
-    let stockIndex = 0,
-      demandIndex = 0;
-
-    while (stockIndex < stocks.length && demandIndex < demands.length) {
-      const currentStock = stocks[stockIndex] || 0,
-        currentDemand = demands[demandIndex] || 0;
-      const transported =
-        currentDemand < currentStock ? currentDemand : currentStock;
-
-      resultTable[stockIndex][demandIndex].transported = transported;
-      demands[demandIndex] = currentDemand - transported;
-      stocks[stockIndex] = currentStock - transported;
-
-      if (currentDemand < currentStock) demandIndex++;
-      else {
-        stockIndex++;
-      }
-
-      this.calculationProcess$.next({
-        transportation: JSON.parse(
-          JSON.stringify(resultTable),
-        ) as TransportTable,
-        demands: [...demands],
-        stocks: [...stocks],
-      });
-    }
-
-    return {
-      epsilon: this.getEpsilon(resultTable),
-      table: resultTable,
-    };
-  }
-
-  public checkSolvability(transportProblemData: TPData): boolean {
-    const {costs, shopDemands, storageStocks} = transportProblemData;
-    if (
-      costs.length === 0 ||
-      shopDemands.length === 0 ||
-      storageStocks.length === 0
-    )
-      return false;
-    const shopDemandSum = sum(shopDemands);
-    const storageStockSum = sum(storageStocks);
-    return shopDemandSum === storageStockSum;
+  public calculateFirstPhase(
+    transportProblemData: TPData,
+    type: TPMethods = 'north-west',
+  ): Observable<Array<CalculationProcess>> {
+    return of(this.firstPhase[type].calculate(transportProblemData));
   }
 
   public getEpsilon(resultTable: TransportTable): number {
@@ -119,14 +44,4 @@ export class TransportProblemService {
 
     return epsilon;
   }
-
-  private createResultTableFrom(costTable: Table): TransportTable {
-    return costTable.map((row: Row) => this.createNewResultRow(row));
-  }
-
-  private createNewResultRow = (costRow: Row): TransportRow =>
-    Object.keys(costRow).reduce((result: TransportRow, key) => {
-      result[key] = {cost: costRow[key]};
-      return result;
-    }, {});
 }