+++ apps/frontend/src/app/pages/transport-problem/utils/tests/solvability.util.spec.ts
@@ -1,0 +1,23 @@
+import {
+  invalidTpDataMock,
+  tpDataFirstMock,
+  tpDataSecondMock,
+} from '../../tests/transport-problem.mock';
+import {checkSolvability} from '../solvability.util';
+
+describe('SolvabilityUtil', () => {
+  it('should be solvable 1', () => {
+    const result = checkSolvability(tpDataFirstMock);
+    expect(result).toEqual(true);
+  });
+
+  it('should be solvable 1', () => {
+    const result = checkSolvability(tpDataSecondMock);
+    expect(result).toEqual(true);
+  });
+
+  it('should not be solvable', () => {
+    const result = checkSolvability(invalidTpDataMock);
+    expect(result).toEqual(false);
+  });
+});